27/05

+++ SysAlmacenVC/src/main/java/pe/edu/upeu/syscenterlife/repositorio/CompraRepository.java
@@ -1,4 +1,3 @@
-
 package pe.edu.upeu.syscenterlife.repositorio;
 
 import org.springframework.data.jpa.repository.JpaRepository;
@@ -6,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import pe.edu.upeu.syscenterlife.modelo.Compra;
 
 @Repository
-public interface CompraRepository extends JpaRepository<Compra, Integer> {
-    // Puedes definir métodos personalizados aquí si es necesario
+public interface CompraRepository extends JpaRepository<Compra, Long> {
+    
 }
